Rename channel.py to kraus.py
