+++ cirq-core/cirq/protocols/__init__.py
@@ -38,7 +38,7 @@ from cirq.protocols.approximate_equality_protocol import (
     approx_eq,
     SupportsApproximateEquality,
 )
-from cirq.protocols.channel import (
+from cirq.protocols.kraus import (
     channel,
     kraus,
     has_channel,
