+++ cirq-core/cirq/protocols/kraus.py
