+++ cirq-core/cirq/protocols/kraus_test.py
